refactor shared client response helpers
